refactor(dobbelOmega): Moved prisma files to node_modules
