--- src/prisma/.gitignore
@@ -1,1 +1,0 @@
-generated
