+++ src/prisma/seeder/src/dobbelOmega/migateArticles.ts
@@ -1,7 +1,7 @@
 import { vevenIdToPnId } from './IdMapper'
 import upsertOrderBasedOnDate from './upsertOrderBasedOnDate'
 import type { PrismaClient as PrismaClientPn } from '@prisma/client'
-import type { PrismaClient as PrismaClientVeven } from '@/generated/veven'
+import type { PrismaClient as PrismaClientVeven } from '@/prisma-dobbel-omega/client'
 import type { IdMapper } from './IdMapper'
 import type { Limits } from './migrationLimits'
