+++ src/prisma/seeder/src/dobbelOmega/migrateEvents.ts
@@ -1,7 +1,7 @@
 import { vevenIdToPnId, type IdMapper } from './IdMapper'
 import upsertOrderBasedOnDate from './upsertOrderBasedOnDate'
 import type { PrismaClient as PrismaClientPn } from '@prisma/client'
-import type { PrismaClient as PrismaClientVeven } from '@/generated/veven'
+import type { PrismaClient as PrismaClientVeven } from '@/prisma-dobbel-omega/client'
 import type { Limits } from './migrationLimits'
 
 export default async function migrateEvents(
