+++ src/prisma/seeder/src/dobbelOmega/migrateImages.ts
@@ -6,7 +6,7 @@ import { writeFile, mkdir } from 'fs/promises'
 import { join } from 'path'
 import type { Limits } from './migrationLimits'
 import type { PrismaClient as PrismaClientPn } from '@prisma/client'
-import type { PrismaClient as PrismaClientVeven } from '@/generated/veven'
+import type { PrismaClient as PrismaClientVeven } from '@/prisma-dobbel-omega/client'
 
 /**
  * This function migrates images from Veven to PN and adds them to the correct image collection
