+++ src/prisma/seeder/src/dobbelOmega/migrateOmbul.ts
@@ -3,7 +3,7 @@ import { v4 as uuid } from 'uuid'
 import { fileURLToPath } from 'url'
 import { dirname, join } from 'path'
 import { writeFile, mkdir } from 'fs/promises'
-import type { PrismaClient as PrismaClientVeven } from '@/generated/veven'
+import type { PrismaClient as PrismaClientVeven } from '@/prisma-dobbel-omega/client'
 import type { PrismaClient as PrismaClientPn } from '@prisma/client'
 import type { IdMapper } from './IdMapper'
 import type { Limits } from './migrationLimits'
