+++ tsconfig.json
@@ -37,7 +37,7 @@
       "@/api/*": ["src/app/api/*"],
       "@/education/*": ["src/actions/education/*", "src/services/education/*", "src/app/education/*"],
       "@/career/*": ["src/actions/career/*", "src/services/career/*", "src/app/career/*"],
-      "@/generated/*": ["src/prisma/generated/*"],
+      "@/prisma-dobbel-omega/*": ["node_modules/.prisma-dobbel-omega/*"],
       "@/seeder/*": ["src/prisma/seeder/*"],
       "@/*": ["./src/*"]
     }
